+++ app/handler/health.py
@@ -1,0 +1,2 @@
+def get_health():
+    return 'OK'
